+++ example/src/components/SignMessage.tsx
@@ -1,43 +1,53 @@
-import { verifyMessage } from 'ethers/lib/utils'
-import { useSignMessage } from 'wagmi'
+import { useEffect, useState } from 'react'
+import { recoverMessageAddress } from 'viem'
+import { type Address, useSignMessage } from 'wagmi'
 
-export const SignMessage = () => {
-  const signMessage = useSignMessage()
+export function SignMessage() {
+  const [recoveredAddress, setRecoveredAddress] = useState<Address>()
+  const {
+    data: signature,
+    variables,
+    error,
+    isLoading,
+    signMessage,
+  } = useSignMessage()
+
+  useEffect(() => {
+    ;(async () => {
+      if (variables?.message && signature) {
+        const recoveredAddress = await recoverMessageAddress({
+          message: variables?.message,
+          signature,
+        })
+        setRecoveredAddress(recoveredAddress)
+      }
+    })()
+  }, [signature, variables?.message])
 
   return (
-    <div>
+    <>
       <form
         onSubmit={(event) => {
           event.preventDefault()
           const element = event.target as HTMLFormElement
           const formData = new FormData(element)
           const message = formData.get('message') as string
-          signMessage.signMessage({ message })
+          signMessage({ message })
         }}
       >
         <input name="message" type="text" required />
-        <button disabled={signMessage.isLoading}>
-          {signMessage.isLoading ? 'Check Wallet' : 'Sign Message'}
+        <button disabled={isLoading} type="submit">
+          {isLoading ? 'Check Wallet' : 'Sign Message'}
         </button>
       </form>
 
-      {signMessage.data && (
+      {signature && (
         <div>
-          <div>signature {signMessage.data}</div>
-          <div>
-            recovered address{' '}
-            {verifyMessage(
-              signMessage.variables?.message as string,
-              signMessage.data,
-            )}
-          </div>
+          <div>Signature: {signature}</div>
+          <div>Recovered address: {recoveredAddress}</div>
         </div>
       )}
-
-      <div>
-        {signMessage.error &&
-          (signMessage.error?.message ?? 'Failed to sign message')}
-      </div>
-    </div>
+      {error && <div>Error: {error?.message}</div>}
+    </>
   )
 }